Move FulfilmentRequest to confirm handler, update formatting on UI and translation
