+++ app/views/handlers/confirmation_email.py
@@ -1,21 +1,15 @@
-from dataclasses import dataclass
 from functools import cached_property
-from typing import Mapping, Optional
+from typing import Optional
 
 from flask import current_app
 from flask_babel import gettext, lazy_gettext
 from itsdangerous import BadSignature
 from werkzeug.exceptions import BadRequest
 
-from app.cloud_tasks.exceptions import CloudTaskCreationFailed
-from app.data_models import FulfilmentRequest, SessionData, SessionStore
+from app.data_models import SessionData, SessionStore
 from app.forms.email_form import EmailForm
 from app.helpers import url_safe_serializer
 from app.questionnaire import QuestionnaireSchema
-from app.settings import (
-    EQ_SUBMISSION_CONFIRMATION_CLOUD_FUNCTION_NAME,
-    EQ_SUBMISSION_CONFIRMATION_QUEUE,
-)
 from app.views.contexts.email_form_context import build_confirmation_email_form_context
 
 
@@ -24,10 +18,6 @@ class ConfirmationEmailNotEnabled(Exception):
 
 
 class ConfirmationEmailLimitReached(Exception):
-    pass
-
-
-class ConfirmationEmailFulfilmentRequestPublicationFailed(Exception):
     pass
 
 
@@ -76,26 +66,6 @@ class ConfirmationEmail:
             return gettext("Error: {page_title}").format(page_title=self.page_title)
         return self.page_title
 
-    def _publish_fulfilment_request(self):
-        fulfilment_request = ConfirmationEmailFulfilmentRequest(
-            self.form.email.data, self._session_store.session_data, self._schema
-        )
-
-        try:
-            return current_app.eq["cloud_tasks"].create_task(
-                body=fulfilment_request.message,
-                queue_name=EQ_SUBMISSION_CONFIRMATION_QUEUE,
-                function_name=EQ_SUBMISSION_CONFIRMATION_CLOUD_FUNCTION_NAME,
-                fulfilment_request_transaction_id=fulfilment_request.transaction_id,
-            )
-        except CloudTaskCreationFailed as exc:
-            raise ConfirmationEmailFulfilmentRequestPublicationFailed from exc
-
-    def handle_post(self):
-        self._publish_fulfilment_request()
-        self._session_store.session_data.confirmation_email_count += 1
-        self._session_store.save()
-
     @staticmethod
     def is_limit_reached(session_data: SessionData) -> bool:
         return (
@@ -106,23 +76,3 @@ class ConfirmationEmail:
     @staticmethod
     def is_enabled(schema: QuestionnaireSchema):
         return schema.get_submission().get("confirmation_email")
-
-
-@dataclass
-class ConfirmationEmailFulfilmentRequest(FulfilmentRequest):
-    email_address: str
-    session_data: SessionData
-    schema: QuestionnaireSchema
-
-    def _payload(self) -> Mapping:
-        return {
-            "fulfilmentRequest": {
-                "email_address": self.email_address,
-                "display_address": self.session_data.display_address,
-                "form_type": self.schema.form_type,
-                "language_code": self.session_data.language_code,
-                "region_code": self.schema.region_code,
-                "questionnaire_id": self.session_data.questionnaire_id,
-                "tx_id": self.session_data.tx_id,
-            }
-        }
